src/main/java/com/github/brunomndantas/jscrapper/support/parser/array/text/reference/ArrayReferenceShortTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.array.text.reference;

import com.github.brunomndantas.jscrapper.support.parser.array.text.ArrayTextParser;
import org.openqa.selenium.WebElement;

public class ArrayReferenceShortTextParser extends ArrayTextParser {

    public ArrayReferenceShortTextParser() {
        super(Short.class, null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Short.parseShort(element.getText());
    }

}