src/test/java/com/github/brunomndantas/jscrapper/support/parser/array/text/reference/ArrayReferenceShortTextParserTest.java
package com.github.brunomndantas.jscrapper.support.parser.array.text.reference;

import com.github.brunomndantas.jscrapper.DummyElement;
import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertSame;

public class ArrayReferenceShortTextParserTest {

    @Test
    public void constructorTest() {
        ArrayReferenceShortTextParser parser = new ArrayReferenceShortTextParser();
        assertSame(Short.class, parser.getKlass());
        assertEquals(null, parser.getDefaultValue());
    }

    @Test
    public void parseElement() throws Exception {
        ArrayReferenceShortTextParser parser = new ArrayReferenceShortTextParser();

        assertEquals((short)2, parser.parseElement(new DummyElement(){
            @Override
            public String getText() {
                return "2";
            }
        }));
    }

}